requirements.txt
CouchDB==0.9
Logbook==0.7.0
PyYAML==3.11
SQLAlchemy==0.9.7
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
coverage==3.7.1
dicttoxml==1.5.5
importlib==1.0.3
inflect==0.2.5
ipdb==0.8
ipython==2.1.0
nose>=1.3.3
nose-exclude==0.2.0
psutil==2.1.1
python-dateutil>=1.5
requests
tornado
wsgiref==0.1.2
xmltodict==0.9.0
PyVCF==0.6.7
